+++ src/test/java/cloud/filibuster/junit/tests/filibuster/server/basic/JUnitFilibusterExternalHttpTest.java
@@ -1,4 +1,4 @@
-package cloud.filibuster.junit.tests.filibuster.server;
+package cloud.filibuster.junit.tests.filibuster.server.basic;
 
 import cloud.filibuster.instrumentation.TestHelper;
 import cloud.filibuster.instrumentation.helpers.Networking;
@@ -22,16 +22,15 @@ import java.util.Arrays;
 import java.util.List;
 
 import static cloud.filibuster.junit.Assertions.wasFaultInjected;
-import static cloud.filibuster.junit.Assertions.wasFaultInjectedOnService;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitFilibusterHttpTest extends JUnitBaseTest {
+public class JUnitFilibusterExternalHttpTest extends JUnitBaseTest {
     private static int numberOfTestsExceptionsThrownFaultsInjected = 0;
 
-    private final List<String> validErrorCodes = Arrays.asList("404", "503");
+    private final List<String> validErrorCodes = Arrays.asList("424", "500");
 
     /**
      * Inject faults between Hello and World using Filibuster and assert proper faults are injected.
@@ -40,11 +39,13 @@ public class JUnitFilibusterHttpTest extends JUnitBaseTest {
     @ExtendWith(GitHubActionsSkipInvocationInterceptor.class)
     @FilibusterTest
     @Order(1)
-    public void testHelloAndWorldServiceWithFilibuster() {
+    public void testHelloAndExternalServiceWithFilibuster() {
+        boolean expected = false;
+
         try {
             String baseURI = "http://" + Networking.getHost("hello") + ":" + Networking.getPort("hello") + "/";
             WebClient webClient = TestHelper.getTestWebClient(baseURI);
-            RequestHeaders getHeaders = RequestHeaders.of(HttpMethod.GET, "/world", HttpHeaderNames.ACCEPT, "application/json");
+            RequestHeaders getHeaders = RequestHeaders.of(HttpMethod.GET, "/external", HttpHeaderNames.ACCEPT, "application/json");
             AggregatedHttpResponse response = webClient.execute(getHeaders).aggregate().join();
             ResponseHeaders headers = response.headers();
             String statusCode = headers.get(HttpHeaderNames.STATUS);
@@ -52,7 +53,6 @@ public class JUnitFilibusterHttpTest extends JUnitBaseTest {
             if (wasFaultInjected()) {
                 numberOfTestsExceptionsThrownFaultsInjected++;
                 assertTrue(validErrorCodes.contains(statusCode));
-                assertTrue(wasFaultInjectedOnService("world"));
             } else {
                 assertEquals("200", statusCode);
             }
