+++ src/test/java/cloud/filibuster/junit/tests/filibuster/server/basic/JUnitFilibusterTestWithNoBeforeEachOrAfterEach.java
@@ -1,4 +1,4 @@
-package cloud.filibuster.junit.tests.filibuster.server;
+package cloud.filibuster.junit.tests.filibuster.server.basic;
 
 import cloud.filibuster.examples.Hello;
 import cloud.filibuster.examples.HelloServiceGrpc;
